Updates to api - split viewsets inot package and implement example giving expected api back for categories
