+++ django_project/api/serializers/point_of_interest.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from trail_mapper.models.point_of_interest import POI
+from trail_mapper.models.point_of_interest import PointOfInterest
 
 __author__ = 'Alison Mukoma <alison@kartoza.com>'
 __date__ = '01/10/2018'
@@ -11,7 +11,7 @@ class POISerializer(serializers.ModelSerializer):
     """Serializer class for a Point of Interest model."""
 
     class Meta:
-        model = POI
+        model = PointOfInterest
         url_field_name = 'image'
         fields = [
             'guid',
