+++ django_project/api/urls.py
@@ -3,7 +3,7 @@
 from django.conf.urls import include, url
 from rest_framework import routers
 
-from views import (
+from viewsets import (
     CategoryViewSet,
     GradeViewSet,
     PointOfInterestViewSet,
