+++ django_project/api/viewsets/trail_section.py
@@ -1,0 +1,12 @@
+from rest_framework import viewsets
+from rest_framework.response import Response
+
+from ..serializers import TrailSectionSerializer
+from trail_mapper.models import TrailSection
+
+
+class TrailSectionViewSet(viewsets.ModelViewSet):
+    queryset = TrailSection.objects.all()
+    serializer_class = TrailSectionSerializer
+    lookup_field = 'slug'
+
